Add my configuration for X
Right now I used KDE's plasma window/desktop manager and with is KDE's
konsole as a terminal emulator. Therefore I don't really need the
.Xresources files which right now just contains a colorsheme, but that
might change in the future, therefore I keep it as a reference.

I use 'startx' to start the X server. It sources .xinitrc and that
starts plasma at the end. Before that I do have some commands that are
commented out. Usually I don't like doing that, but I left it like that,
because there is no prior version of that file in this git repo and I
might go back using qtile and than I'll need these lines and I don't
want to look them up again.