Add my tmuxinator configuration
It just contains a link to a directory in my home folder. I don't want
to have the files in that folder in a public repo. Also in that folder
are configuration files for specific projects (that are used to open
tmux with tmuxinator) and I don't want to commit to this repo just
because I add a new config file for a (temporary) project.