Split alacritty.yml
Now there is a base configuration file, that holds all the configuration except
for the color scheme. The color scheme is moved into separate files. This
was done to be able to write a script that concatenates the base file
with one of the color scheme files into the actual configuration file
for alacritty. This way I want to be able to switch between color
schemes with a key combination.