Change the set_colorscheme script
The scipt works a little different now. Instead of concactenating
different files into the actual config file, now the config file
contains all available color scheme and the current one is set in a
single line.

The advantage is, that in case I want to change other configurations
(not the colorsheme), I can just change the alacritty.yml file and
everythnig will work just fine. Before I would have had to change the
base.yml file and for the changes to take effect, I would have had to
make a color scheme switch. This is confusing and I'm pretty sure, it
would lead to confusions some time in the future.