Turn numlock on when the X-server is started
In my .bash_profile file I have a line 'setleds -D +num' which enables
the numlock when a login shell invokes bash. So basically after logging
in into my machine, numlock is activated. Unfortunately when I start the
X-server, the LED on my numlock key is still on, but
the number key block has the alternative keys and not the number
activated. Therefore I installed numlockx and call if from within
.xinitrc

I might actually move this line into .xinitrc, but for now it works.