Decrease update interval of battery widget
When plugging the power supply into the laptop or unplugging it, it was
a littel annoying that it took so long until the display of the battery
status was updated. I figured 5 seconds was a not too long to wait in
such a case. I didn't want to make it even shorter, because acually the
battery status shouldn't change very often and therefore updating the
widget more often does not seem to make sense at all.