Add first version of a install config file
This file is supposed to hold all the configs that I want to install
into my system. It list all directories and files for which links or
copies have to be put into other places. The idea is to read that file
in with a script and create all the links and copies.