Move everything in etc into a separate config file
The etc stuff requires root privileges and this causes problems. Because
if you run the install.py script with root privileges the part that
involves /etc works, but everything else is linked/copied into the home
directory of the root user. And thats not necessarily what I want. On
the other hand, if you run the script without root privileges, than the
etc stuff does not work. I have to find a solution for it. I kind of
don't like how it works right now.