Add navigation between vim and tmux
I used to have these mappings when I used tmux (basically as a window
manager). I removed them, when I installed qtile and didn't need them
anymore. And they caused problems when vim was started outside of tmux.
Therefore I had removed tham and now re-added them with a slight
modification (check if vim is run in tmux or not).