Fix bug: make sure to call all sinks
I had the problem that in the output of 'pactl list sinks short' the
sind number doesn't necessarily have to be just one byte. I notices it
was '10' for example. In such a case the flag '-b1' for the 'cut'
command, didn't work as expected. In the case of '10' it just returns
'1' which is apparently work. In such a case the whole command to change
the volume didn't work as intendet.