Adjust links to bookmarks and quickmarks
Qutebrowser saves bookmarks in a folder in
XDG_CONFIG_HOME/qutebrowser/bookmarks. I linked that folder to a folder
in my home dirctory called ~/workspaces/qutebrowser/bookmarks. Since I
want to get rid of the workspace directory, I had to adjust the link for
the bookmarks. The same is true for the quickmarks (execpt it's a file
not a folder in this case).