Add links that are needed to build xmonad
I did want to build xmonad from the source files, because the version
that is currently packaged in the Arch repos is a little behind the
current version of xmonad. Since I needed some features of the newest
available version, I had to build it. But I didn't want the source code
of xmonad im my config folder. So I just added links.