Show workspace set one on startup
This was done to have the right order of workspaces on startup. Without
the change my workspace 2 was shown on my left most monitor and
workspace 1 on my middle monitor on my three monitor setup. This is due
to the fact that I run a xrandr script on startup that places the
prmariy monitor in the middle.