Replace call to "hostname"
I realized on my second laptop that hostname is not part of the standard
Arch installation. Therefore I had to install it with the iutils package
I think. Anyway the hostname can be fount in /etc/hostname. No need to
install any additional packages for the startup script to run correctly.