[xmonad] Make sure last window focus is kept
I use KeePassXC to handle my passwords. When I wanted to enter a
password in the terminal and used KeePassXC's auto type short cut, the
password was not entered in the right terminal window after the auto
type window closed. At least not if there were multiple windows on the
workspace. This was pretty annoying. This commit fixes this issue.