[urlscan] Copy "colorized" to "default" pallet
I wanted the colorized pallet to be the one that is used when `urlscan`
is invoked. Somehow moving it to the top didn't help. It also didn't
help to just remove the "default" pallet. It seems as if urlscan uses
the "default" pallet either way if it is explicitly defined in the
config file or not.