[nvim] Re-add scrolloff option (with fewer lines)
I had added this option before, but with 999 lines. This caused the
cursor to stay in the middle of the screen, which kind of annoyed me, so
I deleted the option a while back. But I also don't like to move the
currently used line all the way to the top or button. Having five lines
above and below seems to be a good compromise.